ObjSeq_ApplyFrameCurves scroll: float->s16 direct conv (drop (int) cast) stores raw fctiwz int to tex1->offsetS w/o canonicalizing extsh, tex2 keeps neg+extsh via (s16)-scroll (99.765->99.933, unit 99.464->99.477)

+++ src/main/objseq.c
@@ -2300,7 +2300,7 @@ void ObjSeq_ApplyFrameCurves(u8* obj, u8* seqObj, u8* seq, int frame)
                 {
                     val = lbl_803DEFB0;
                 }
-                scroll = (int)(gObjSeqTexScrollScale * val);
+                scroll = gObjSeqTexScrollScale * val;
                 if (tex1 != NULL)
                 {
                     tex1->offsetS = scroll;
